+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   * [FIX]: Ad card option modals properly styled
   * The custom splash stylesheet for each publisher is now included when
     in the editor
+  * [FEATURE]: The splash ratio/theme of your last MiniReel is now
+    used as the default when you create a new MiniReel
   * The embed code will now be configured for preloading on inline
     display modes
     * **v1.1.0 of c6embed must be deployed for this to work**